correcciones varias y listo
