--- app/src/main/java/com/example/ongsomosmas/Fragment/LogInViewModel.kt
@@ -1,35 +1,0 @@
-package com.example.ongsomosmas.Fragment
-
-import androidx.core.util.PatternsCompat
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import java.util.regex.Pattern
-
-class LogInViewModel : ViewModel() {
-
-    val enableButton = MutableLiveData<Boolean>(false)
-
-    private fun validateEmail(email: String): Boolean {
-        return if (email.isEmpty()) {
-            false
-        } else PatternsCompat.EMAIL_ADDRESS.matcher(email).matches()
-    }
-
-    private fun validatePassword(password: String): Boolean {
-        val passwordRegex = Pattern.compile(
-            "^" +
-                    "(?=.*[0-9])" +          //al menos 1 dígito
-                    "(?=.*[A-Z])" +          //al menos una minúscula
-                    ".{8,}" +                //minimo 8 de largo
-                    "$"
-        )
-        return if (password.isEmpty()) {
-            false
-        } else passwordRegex.matcher(password).matches()
-    }
-
-    fun validate(email: String, password: String){
-        val respuesta = validateEmail(email) && validatePassword(password)
-        enableButton.value = !respuesta
-    }
-}
